backend/graphql_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
supabase
python-jose[cryptography]
strawberry-graphql[fastapi]